requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ultralytics>=8.0.0  # For YOLO
git+https://github.com/openai/CLIP.git  # For CLIP
segment-anything>=1.0  # For SAM
yt-dlp>=2023.3.4  # For YouTube video download
tqdm>=4.65.0  # For progress bars
requests>=2.31.0
matplotlib
shapely
moviepy
supervision
python-dotenv==1.0.0
transformers>=4.36.2  # For Florence-2
einops>=0.7.0  # Required by Florence-2
